refactor: hooks/api => hooks/apis
