+++ dashboard/src/hooks/apis/devices.hook.ts
